Move listing search function into utils
Also simplify/flatten settings.
